adds message builder

+++ Data/Repositories/ChecklistMessageRepository.cs
@@ -12,10 +12,11 @@ public class ChecklistMessageRepository : IChecklistMessageRepository
         _context = context;
     }
 
-    public ChecklistMessage AddOrUpdate(long chatId, long messageId, long checklistId)
+    public ChecklistMessage AddOrUpdate(long chatId, int messageId, long checklistId)
     {
         var checklistMessage = Get(chatId, checklistId);
-        if (checklistMessage is null) {
+        if (checklistMessage is null)
+        {
             checklistMessage = new ChecklistMessage
             {
                 ChatId = chatId,
@@ -23,7 +24,9 @@ public class ChecklistMessageRepository : IChecklistMessageRepository
                 MessageId = messageId
             };
             _context.ChecklistMessages.Add(checklistMessage);
-        } else {
+        }
+        else
+        {
             checklistMessage.MessageId = messageId;
         }
         _context.SaveChanges();
