+++ Data/Repositories/IChecklistMessageRepository.cs
@@ -2,9 +2,10 @@ using AnotherChecklistBot.Models;
 
 namespace AnotherChecklistBot.Data.Repositories;
 
-public interface IChecklistMessageRepository {
+public interface IChecklistMessageRepository
+{
     public ChecklistMessage? Get(long chatId, long checklistId);
 
-    public ChecklistMessage AddOrUpdate(long chatId, long messageId, long checklistId);
-    
+    public ChecklistMessage AddOrUpdate(long chatId, int messageId, long checklistId);
+
 }
