+++ Data/Repositories/IChecklistRepository.cs
@@ -2,7 +2,8 @@ using AnotherChecklistBot.Models;
 
 namespace AnotherChecklistBot.Data.Repositories;
 
-public interface IChecklistRepository {
-    public Checklist Create(long sourceChatId, long sourceMessageId, ICollection<string> items);
+public interface IChecklistRepository
+{
+    public Checklist Create(long sourceChatId, int sourceMessageId, ICollection<string> items);
     public Checklist? GetById(long id);
 }
